# data/fossology-LicenseDb/cmd/laas/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Service to host license information for other services to query over REST API.",
        "title": "laas (License as a Service) API",
        "contact": {
            "name": "FOSSology",
            "url": "https://fossology.org",
            "email": "fossology@fossology.org"
        },
        "license": {
            "name": "GPL-2.0-only",
            "url": "https://github.com/fossology/LicenseDb/blob/main/LICENSE"
        },
        "version": "0.0.9"
    },
    "basePath": "/api/v1",
    "paths": {
        "/apiCollection": {
            "get": {
                "description": "Returns the apis which require authentication and which do not",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API Collection"
                ],
                "summary": "Returns the apis which require authentication and which do not",
                "operationId": "getAPICollection",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.APICollectionResponse"
                        }
                    },
                    "500": {
                        "description": "Unable to parse swagger docs",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    }
                }
            }
        },
        "/audits": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "{}": []
                    }
                ],
                "description": "Get all audit records from the server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audits"
                ],
                "summary": "Get audit records",
                "operationId": "GetAllAudit",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of records per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Audit records",
                        "schema": {
                            "$ref": "#/definitions/models.AuditResponse"
                        }
                    },
                    "404": {
                        "description": "Not changelogs in DB",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    }
                }
            }
        },
        "/audits/{audit_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "{}": []
                    }
                ],
                "description": "Get a specific audit records by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audits"
                ],
                "summary": "Get an audit record",
                "operationId": "GetAudit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Audit ID",
                        "name": "audit_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AuditResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid audit ID",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    },
                    "404": {
                        "description": "No audit entry with given ID",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    }
                }
            }
        },
        "/audits/{audit_id}/changes": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "{}": []
                    }
                ],
                "description": "Get changelogs of an audit record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audits"
                ],
                "summary": "Get changelogs",
                "operationId": "GetChangeLogs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Audit ID",
                        "name": "audit_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ChangeLogResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid audit ID",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    },
                    "404": {
                        "description": "No audit entry with given ID",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    },
                    "500": {
                        "description": "unable to find changes",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    }
                }
            }
        },
        "/audits/{audit_id}/changes/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "{}": []
                    }
                ],
                "description": "Get a specific changelog of an audit record by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audits"
                ],
                "summary": "Get a changelog",
                "operationId": "GetChangeLogbyId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Audit ID",
                        "name": "audit_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Changelog ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ChangeLogResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    },
                    "404": {
                        "description": "No changelog with given ID found",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    }
                }
            }
        },
        "/dashboard": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "{}": []
                    }
                ],
                "description": "Fetches data to be displayed on the dashboard",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Fetches data to be displayed on the dashboard",
                "operationId": "GetDashboardData",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DashboardResponse"
                        }
                    },
                    "500": {
                        "description": "Something went wrong",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    }
                }
            }
        },
...
        "/licenses": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "{}": []
                    }
                ],
                "description": "Filter licenses based on different parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Licenses"
                ],
                "summary": "Filter licenses",
                "operationId": "FilterLicense",
                "parameters": [
                    {
                        "type": "string",
                        "description": "SPDX ID of the license",
                        "name": "spdxid",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Active license only",
                        "name": "active",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "OSI Approved flag status of license",
                        "name": "osiapproved",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Copyleft flag status of license",
                        "name": "copyleft",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit of responses per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "External reference parameters",
                        "name": "externalRef",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "spdx_id",
                            "shortname",
                            "fullname"
                        ],
                        "type": "string",
                        "default": "shortname",
                        "description": "Sort by field",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "asc",
                        "description": "Asc or desc ordering",
                        "name": "order_by",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Filtered licenses",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid value",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new license in the service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Licenses"
                ],
                "summary": "Create a new license",
                "operationId": "CreateLicense",
                "parameters": [
                    {
                        "description": "New license to be created",
                        "name": "license",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LicenseCreateDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "New license created successfully",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    },
                    "500": {
                        "description": "Failed to create license",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    }
                }
            }
        },
        "/licenses/export": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "{}": []
                    }
                ],
                "description": "Export all licenses as a json file",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Licenses"
                ],
                "summary": "Export all licenses as a json file",
                "operationId": "ExportLicenses",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.LicenseResponseDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch Licenses",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    }
                }
            }
        },
...
        "/licenses/preview": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": [],
                        "{}": []
                    }
                ],
                "description": "Get shortnames and ids of all active licenses from the service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Licenses"
                ],
                "summary": "Get shortnames and ids of all active licenses",
                "operationId": "GetAllLicensePreviews",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Active license only",
                        "name": "active",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.LicensePreviewResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid active value",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    },
                    "500": {
                        "description": "Unable to fetch licenses",
                        "schema": {
                            "$ref": "#/definitions/models.LicenseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.APICollectionResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.APICollection"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
...
        "models.AuditResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Audit"
                    }
                },
                "paginationmeta": {
                    "$ref": "#/definitions/models.PaginationMeta"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
...
        "models.LicenseError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid request body"
                },
                "message": {
                    "type": "string",
                    "example": "invalid request body"
                },
                "path": {
                    "type": "string",
                    "example": "/api/v1/licenses"
                },
                "status": {
                    "type": "integer",
                    "example": 400
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-12-01T10:00:51+05:30"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Token from /login endpoint. Enter the token with the `Bearer ` prefix, e.g. \\\"Bearer eyJhbGciOiJ.....\\\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}